requirements.txt
astroid==2.2.5
Django==2.2.5
django-friendship==1.8.1
lazy-object-proxy==1.4.2
pylint==2.3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
